+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.7.0] - (22.02.2021)
+## Changes
+- code rebuild, use Characteristic.onSet/onGet
+- require Homebridge 1.3.x or above
 
 ## [1.6.6] - (06.01.2021)
 ## Changs